Hide sea super-regions
After Paradox added sea super-regions, option to hide no-superregion provinces only affects wastelands. It was necessary to add feature that would hide sea-provinces as well. By default original sea-superregions are hidden but the list of them can be costumized in the Configurator. This feature can as well be disabled.
Other changes:
- Last element of assignment file is no longer ignored
- Better logging of launch time (program no longer presumpts the length of number)